requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
meraki==1.30.0
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.14
webex-bot==0.4.0
webexteamsbot==0.1.4.2
webexteamssdk==1.6.1
websockets==10.2
Werkzeug==2.2.3
yarl==1.8.2
